requirements.txt
Django==2.2.4
djangorestframework==3.9.4
pytz==2019.2
sqlparse==0.3.0
cryptography==2.7
pycryptodome
